% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_stats.R
\name{format_stats.BFBayesFactor}
\alias{format_stats.BFBayesFactor}
\title{Format Bayes factors}
\usage{
\method{format_stats}{BFBayesFactor}(x, ...)
}
\description{
Format Bayes factors
}
\seealso{
Other functions for printing statistical objects: 
\code{\link{format_bf}()},
\code{\link{format_corr}()},
\code{\link{format_stats}()},
\code{\link{format_stats.htest}()},
\code{\link{format_ttest}()}
}
\concept{functions for printing statistical objects}
