% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_stats.R
\name{format_ttest}
\alias{format_ttest}
\title{Format t-test statistics}
\usage{
format_ttest(x, digits, pdigits, pzero, full, italics, dfs, mean, type)
}
\arguments{
\item{x}{An \code{htest} object}

\item{digits}{Number of digits after the decimal for means, confidence
intervals, and test statistics}

\item{pdigits}{Number of digits after the decimal for p-values, ranging
between 1-5 (also controls cutoff for small p-values)}

\item{pzero}{Logical value (default = FALSE) for whether to include
leading zero for p-values}

\item{full}{Logical value (default = TRUE) for whether to include means
and confidence intervals or just test statistic and p-value}

\item{italics}{Logical value (default = TRUE) for whether \emph{p} label should be
italicized}

\item{dfs}{Formatting for degrees of freedom ("par" = parenthetical,
"sub" = subscript, "none" = do not print degrees of freedom)}

\item{mean}{Formatting for mean label ("abbr" = M, "word" = Mean)}

\item{type}{Type of formatting ("md" = markdown, "latex" = LaTeX)}
}
\value{
A character string of statistical information formatted in Markdown or LaTeX.
}
\description{
With \code{format_ttest()} you can format t-tests generated from \code{t.test()} and
\code{wilcox.test()} output. The default output is APA formatted, but numbers of
digits, leading zeros, the presence of means and confidence intervals,
italics, degrees of freedom, and mean labels are all customizable.
}
\examples{
# Prepare data
mtcars_tt <- t.test(formula = mtcars$mpg ~ mtcars$vs)
# Print statistics
format_stats(mtcars_tt)
# Change digits
format_stats(mtcars_tt, digits = 2)
# Change cutoff digits for p-value
format_stats(mtcars_tt, pdigits = 2)
# Add leading zero to p-value and don't print confidence intervals
format_stats(mtcars_tt, pzero = TRUE, full = FALSE)
}
\seealso{
Other functions for printing statistical objects: 
\code{\link{format_bf}()},
\code{\link{format_corr}()},
\code{\link{format_stats}()},
\code{\link{format_stats.BFBayesFactor}()},
\code{\link{format_stats.htest}()}
}
\concept{functions for printing statistical objects}
