% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_stats.R
\name{format_stats.htest}
\alias{format_stats.htest}
\title{Format hypothesis test statistics}
\usage{
\method{format_stats}{htest}(
  x,
  digits = NULL,
  pdigits = 3,
  pzero = FALSE,
  full = TRUE,
  italics = TRUE,
  dfs = "par",
  mean = "abbr",
  type = "md"
)
}
\arguments{
\item{x}{t-test object}

\item{digits}{Number of digits after the decimal for means, confidence
intervals, and test statistics}

\item{pdigits}{Number of digits after the decimal for p-values, ranging
between 1-5 (also controls cutoff for small p-values)}

\item{pzero}{Logical value (default = FALSE) for whether to include
leading zero for p-values}

\item{full}{Logical value (default = TRUE) for whether to include means
and confidence intervals or just test statistic and p-value}

\item{italics}{Logical value (default = TRUE) for whether \emph{p} label should be
italicized}

\item{dfs}{Formatting for degrees of freedom ("par" = parenthetical,
"sub" = subscript, "none" = do not print degrees of freedom)}

\item{mean}{Formatting for mean label ("abbr" = M, "word" = Mean)}

\item{type}{Type of formatting ("md" = markdown, "latex" = LaTeX)}
}
\description{
Format hypothesis test statistics
}
\details{
\subsection{Correlations}{

With \code{format_corr()} you can format correlation statistics generated from
\code{cor.test()} output. This detects whether the object is from a Pearson,
Spearman, or Kendall correlation and reports the appropriate correlation
label (r, \eqn{\tau}, \eqn{\rho}). The default output is APA formatted, but numbers of digits,
leading zeros, the presence of confidence intervals, and italics are all
customizable.
}

\subsection{T-tests}{

With \code{format_ttest()} you can format t-tests generated from \code{t.test()} and
\code{wilcox.test()} output. The default output is APA formatted, but numbers of
digits, leading zeros, the presence of means and confidence intervals,
italics, degrees of freedom, and mean labels are all customizable.
}
}
\examples{
format_stats(cor.test(mtcars$mpg, mtcars$cyl))
format_stats(t.test(mtcars$vs, mtcars$am))
}
\seealso{
Other functions for printing statistical objects: 
\code{\link{format_bf}()},
\code{\link{format_corr}()},
\code{\link{format_stats}()},
\code{\link{format_stats.BFBayesFactor}()},
\code{\link{format_ttest}()}
}
\concept{functions for printing statistical objects}
